+++ python/sglang/bench_serving.py
@@ -565,6 +565,14 @@ async def benchmark(
         output_len=test_output_len,
         extra_request_body=extra_request_body,
     )
+    test_output = await request_func(request_func_input=test_input)
+    if not test_output.success:
+        raise ValueError(
+            "Initial test run failed - Please make sure benchmark arguments "
+            f"are correctly specified. Error: {test_output.error}"
+        )
+    else:
+        print("Initial test run completed. Starting main benchmark run...")
 
     pbar = None if disable_tqdm else tqdm(total=len(input_requests))
